front-end/requirements.txt
fastapi~=0.82.0
google-cloud-tasks~=2.7
uvicorn~=0.18.3
pydantic~=1.10.2
loguru~=0.6.0
redis~=4.3.4